Removed some readmes, some refactoring
