+++ package.json
@@ -20,12 +20,12 @@
     "@roast-cms/react-sugar-styled": "^1.0.4",
     "axios": "0.18.0",
     "immutable": "^3.8.2",
-    "keycode": "^2.1.9",
+    "keycode": "^2.2.0",
     "localforage": "^1.6.0",
     "localforage-getitems": "^1.4.1",
     "lodash": "^4.17.4",
     "oauth-open": "^1.0.3",
-    "query-string": "^5.0.1",
+    "query-string": "5.1.1",
     "react": "^16.0.0",
     "react-dom": "^16.0.0",
     "react-ga": "^2.4.0",
@@ -43,7 +43,7 @@
     "typeface-exo-2": "^0.0.54",
     "typeface-lora": "^0.0.54",
     "uuid": "^3.1.0",
-    "webpack": "^4.1.1"
+    "webpack": "4.2.0"
   },
   "devDependencies": {
     "babel-plugin-root-import": "^5.1.0",
