--- src/app/components/_controls/README.md
@@ -1,4 +1,0 @@
-# Controls
-Components that resemble user controls live here.
-
-By default this folder contains pure functions (components). If closely-related container (class component) needs to live close-by, it'll be in the /containers folder.
