--- src/app/components/_icons/README.md
@@ -1,4 +1,0 @@
-# Icons
-Graphics, specifically as images or components reside in this folder.
-
-By default this folder contains pure functions (components). If closely-related container (class component) needs to live close-by, it'll be in the /containers folder.
