--- src/app/components/_rt-snippets/README.md
@@ -1,4 +1,0 @@
-# Rich Text snippets
-Whenever some rich text (could be or not be a complete component, hence "snippets") needs to be reused/refactored this is the folder to keep them in.
-
-By default this folder contains pure functions (components). If closely-related container (class component) needs to live close-by, it'll be in the /containers folder.
