--- src/app/components/_screens/README.md
@@ -1,4 +1,0 @@
-# Screens
-Components resembling screens or views live here.
-
-By default this folder contains pure functions (components). If closely-related container (class component) needs to live close-by, it'll be in the /containers folder.
