--- src/app/containers/_forms/README.md
@@ -1,4 +1,0 @@
-# Forms
-Components necessary to create forms reside here.
-
-By default this folder contains container components (React classes) or pure functions requiring Redux connections. If closely-related pure functions (components) needs to live close-by, it'll be in the /containers folder.
