--- src/app/containers/_screens-auth/README.md
@@ -1,4 +1,0 @@
-# Auth Screens
-Views or screens necessary to create experience that requires user login or authorized access live here.
-
-By default this folder contains container components (React classes) or pure functions requiring Redux connections. If closely-related pure functions (components) needs to live close-by, it'll be in the /containers folder.
