--- src/app/containers/_screens-errors/README.md
@@ -1,4 +1,0 @@
-# Error screens
-This folder hosts components that determine error state and show it to the user.
-
-By default this folder contains container components (React classes) or pure functions requiring Redux connections. If closely-related pure functions (components) needs to live close-by, it'll be in the /containers folder.
